+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.15.1
+  * Added a utility to convert examples in calls to be const. This is important
+    for DDC usage, where they aren't optimized away.
+
 ## 0.15.0
   * Change non-transformer message rewriting to preserve the original message as
     much as possible. Adds --useStringSubstitution command-line arg.